pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apollobot"
version = "0.1.0"
description = "ApolloBot — the autonomous research engine by Frontier Science. Give it a question, get back a paper."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Frontier Science Contributors" },
]
keywords = ["research", "ai", "agent", "science", "mcp", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "textual>=0.80.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "statsmodels>=0.14.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "biopython>=1.83",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
latex = [
    "pylatex>=1.4.2",
]
gpu = [
    "torch>=2.2.0",
    "transformers>=4.38.0",
]
notifications = [
    "httpx>=0.27",
]
telegram = [
    "httpx>=0.27",
]
discord = [
    "httpx>=0.27",
]
all-channels = [
    "httpx>=0.27",
]
all = [
    "apollobot[dev,latex,gpu,notifications]",
]

[project.scripts]
apollo = "apollobot.cli:main"

[project.urls]
Homepage = "https://frontierscience.ai"
Documentation = "https://apollobot.dev"
Repository = "https://github.com/frontier-science/apollobot"
Issues = "https://github.com/frontier-science/apollobot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/apollobot"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
